Fix auth on remaining controllers

+++ src/main/java/com/ivanolmo/kanbantaskmanager/service/ColumnServiceImpl.java
@@ -1,41 +1,50 @@
 package com.ivanolmo.kanbantaskmanager.service;
 
+import com.ivanolmo.kanbantaskmanager.dto.BoardInfo;
 import com.ivanolmo.kanbantaskmanager.dto.ColumnDTO;
+import com.ivanolmo.kanbantaskmanager.dto.ColumnInfo;
 import com.ivanolmo.kanbantaskmanager.entity.Board;
 import com.ivanolmo.kanbantaskmanager.entity.Column;
+import com.ivanolmo.kanbantaskmanager.entity.User;
 import com.ivanolmo.kanbantaskmanager.exception.EntityOperationException;
 import com.ivanolmo.kanbantaskmanager.mapper.ColumnMapper;
 import com.ivanolmo.kanbantaskmanager.repository.BoardRepository;
 import com.ivanolmo.kanbantaskmanager.repository.ColumnRepository;
+import com.ivanolmo.kanbantaskmanager.util.UserHelper;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
+@RequiredArgsConstructor
 @Slf4j
 public class ColumnServiceImpl implements ColumnService {
   private final ColumnRepository columnRepository;
   private final BoardRepository boardRepository;
   private final ColumnMapper columnMapper;
-
-  public ColumnServiceImpl(ColumnRepository columnRepository,
-                           BoardRepository boardRepository,
-                           ColumnMapper columnMapper) {
-    this.columnRepository = columnRepository;
-    this.boardRepository = boardRepository;
-    this.columnMapper = columnMapper;
-  }
+  private final UserHelper userHelper;
 
   // create board column
   @Transactional
   public ColumnDTO addColumnToBoard(String boardId, ColumnDTO columnDTO) {
-    // get board, throw error if not found
-    Board board = boardRepository.findById(boardId)
+    // get user from security context via helper method
+    User user = userHelper.getCurrentUser();
+
+    // get board and user info
+    BoardInfo boardInfo = boardRepository.findBoardInfoById(boardId)
         .orElseThrow(() -> new EntityOperationException("Board", "read", HttpStatus.NOT_FOUND));
 
-    // if new column name already exists for this board, throw error
+    // check that column -> board relation and user id matches
+    if (!boardInfo.getUserId().equals(user.getId())) {
+      throw new EntityOperationException(
+          "You do not have permission to add a column to this board", HttpStatus.FORBIDDEN);
+    }
+
+    Board board = boardInfo.getBoard();
+
+    // check if new column name matches an existing name
     columnRepository.findByNameAndBoardId(columnDTO.getName(), boardId)
         .ifPresent(existingColumn -> {
           throw new EntityOperationException("A column with that name already exists",
@@ -60,33 +69,36 @@ public class ColumnServiceImpl implements ColumnService {
   // update column
   @Transactional
   public ColumnDTO updateColumnName(String id, ColumnDTO columnDTO) {
-    // get column by id or else throw exception
-    Column column = columnRepository.findById(id)
+    // Get user from security context via helper method
+    User user = userHelper.getCurrentUser();
+
+    // get column and user info
+    ColumnInfo columnInfo = columnRepository.findColumnInfoById(id)
         .orElseThrow(() -> new EntityOperationException("Column", "read", HttpStatus.NOT_FOUND));
 
-    // get board that this column belongs to or else throw exception
-    // an error being thrown here would signify a data integrity issue
-    Board board = column.getBoard();
-
-    if (board == null) {
-      throw new EntityOperationException("Board", "read", HttpStatus.INTERNAL_SERVER_ERROR);
+    // check if column info and user id matches
+    if (!columnInfo.getUserId().equals(user.getId())) {
+      throw new EntityOperationException(
+          "You do not have permission to update a column in this board", HttpStatus.FORBIDDEN);
     }
 
-    // if column name already exists for this board, throw error
-    columnRepository.findByNameAndBoardId(columnDTO.getName(), board.getId())
+    Column column = columnInfo.getColumn();
+
+    // check if updated name matches an existing name
+    columnRepository.findByNameAndBoardId(columnDTO.getName(), column.getBoard().getId())
         .ifPresent(existingColumn -> {
-          throw new EntityOperationException("A column with that name already exists",
-              HttpStatus.CONFLICT);
+          throw new EntityOperationException(
+              "A column with that name already exists", HttpStatus.CONFLICT);
         });
 
-    // perform update and return dto
+    // update column name and return DTO
     try {
       column.setName(columnDTO.getName());
       Column updatedColumn = columnRepository.save(column);
       return columnMapper.toDTO(updatedColumn);
     } catch (Exception e) {
-      log.error("An error occurred while updating column '{}': {}",
-          columnDTO.getName(), e.getMessage());
+      log.error(
+          "An error occurred while updating column '{}': {}", columnDTO.getName(), e.getMessage());
       throw new EntityOperationException("Column", "update", e, HttpStatus.INTERNAL_SERVER_ERROR);
     }
   }
@@ -94,13 +106,22 @@ public class ColumnServiceImpl implements ColumnService {
   // delete column
   @Transactional
   public void deleteColumn(String id) {
-    // delete column or throw error if column not found
+    // get user from security context via helper method
+    User user = userHelper.getCurrentUser();
+
+    // get column and user info
+    ColumnInfo columnInfo = columnRepository.findColumnInfoById(id)
+        .orElseThrow(() -> new EntityOperationException("Column", "read", HttpStatus.NOT_FOUND));
+
+    // check if column info and user id matches
+    if (!columnInfo.getUserId().equals(user.getId())) {
+      throw new EntityOperationException(
+          "You do not have permission to delete a column from this board", HttpStatus.FORBIDDEN);
+    }
+
+    // delete column
     try {
       columnRepository.deleteById(id);
-    } catch (EmptyResultDataAccessException e) {
-      log.error("An error occurred while deleting column id '{}': {}",
-          id, e.getMessage());
-      throw new EntityOperationException("Column", "delete", HttpStatus.NOT_FOUND);
     } catch (Exception e) {
       log.error("An error occurred while deleting column id '{}': {}",
           id, e.getMessage());
